CMAKE_OSX_ARCHITECTURES must be set with FORCE
`CMAKE_OSX_ARCHITECTURES` is actually assigned on the command line, and if not there, is assigned by CMake to an empty string.  This means that it has a value when the `set` command is encountered, and therefore unless `FORCE` is also specified, the `set` will be ignored.

See #3 for more information.