Add support for duration values in configuration
It would be nice to be able to serialize a duration directly.  Use the count on the duration field proper rather than trying to convert to a universal type; the string representation may have to face the user, and if the developer represents duration with `std::chrono::seconds`, the user should be specifying that duration in seconds as well.