Fix static assert in get_marshaller
This static assert is never actually going to get hit, because the expression `marshaller_base& x = get_marshaller()` is evaluated before the template itself is instantiated.  This means that the compiler will complain about a conversion being impossible and never actually try to instantiate the template.  We have to make the return type strictly compatible and _then_ we can do the check inside the static_assert.